udev: Rework tracepoints using usdt.h
usdt.h is a single-header library from the developers of libbpf to
make using usdt easier (github.com/libbpf/usdt). As the header is
specifically intended to be vendored and there's nothing in there
that could ever be part of a shared library, we vendor the header
inside of our repo (it's only 500 lines anyway).

We immediately make use of the new header by re-implementing udev's
tracepoints on top of usdt.h. Because usdt.h is now always available,
we check if the usdt is currently attached to before generating its
arguments to make sure we don't do unnecessary work if nothing is
attached to the usdt.